refactoring of configuration builder
