+++ src/main/java/com/kryptokrauts/aeternity/sdk/service/keypair/KeyPairServiceConfiguration.java
@@ -1,14 +1,13 @@
 package com.kryptokrauts.aeternity.sdk.service.keypair;
 
-import com.kryptokrauts.aeternity.sdk.service.config.ServiceConfiguration;
+import com.kryptokrauts.aeternity.sdk.service.ServiceConfiguration;
 
 import lombok.Builder.Default;
 import lombok.Getter;
 import lombok.experimental.SuperBuilder;
 
 @Getter
-// @Builder( builderMethodName = "configure", buildMethodName = "compile" )
-@SuperBuilder
+@SuperBuilder( builderMethodName = "configure", buildMethodName = "compile" )
 public class KeyPairServiceConfiguration extends ServiceConfiguration {
     @Default
     private String cipherAlgorithm = "AES/ECB/NoPadding";
